My pydev project
This leaves the framework for a pretty cool project. Let me know if you
don’t understand anything going on here. I put in some files I have
already made, initialized the pygame setting, and made a LOT of classes
for the actual computation behind the game. Make sure you have pygame
installed for this to work. If not, you can always work on the class
file, and add some methods for the Rabbit class